requirements.txt
spotipy
youtube-search
pytube3